feat(backend auth): add middleware for authorization and authentication
